djembe-mvp/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
redis==5.0.1
psycopg2-binary==2.9.9